#16 Create a banner, Made a banner
